fix(domain): fix usecases names
