--- rick-and-morty-app/src/domain/usecases/filter-character-by-name.ts
@@ -1,12 +1,0 @@
-import { CharacterResponseModel } from '../models'
-
-export interface FilterCharacterByName {
-  perform(params: FilterCharacterByName.Params): Promise<FilterCharacterByName.Result>
-}
-
-export namespace FilterCharacterByName {
-  export type Params = {
-    name: string
-  }
-  export type Result = CharacterResponseModel
-}
