--- rick-and-morty-app/src/domain/usecases/filter-character-by-species.ts
@@ -1,12 +1,0 @@
-import { CharacterResponseModel } from '../models'
-
-export interface FilterCharacterBySpecies {
-  perform(params: FilterCharacterBySpecies.Params): Promise<FilterCharacterBySpecies.Result>
-}
-
-export namespace FilterCharacterBySpecies {
-  export type Params = {
-    species: string
-  }
-  export type Result = CharacterResponseModel
-}
