+++ rick-and-morty-app/src/domain/usecases/filter-characters-by-name.ts
@@ -1,0 +1,12 @@
+import { CharacterResponseModel } from '../models'
+
+export interface FilterCharactersByName {
+  perform(params: FilterCharactersByName.Params): Promise<FilterCharactersByName.Result>
+}
+
+export namespace FilterCharactersByName {
+  export type Params = {
+    name: string
+  }
+  export type Result = CharacterResponseModel
+}
