+++ rick-and-morty-app/src/domain/usecases/filter-characters-by-status.ts
@@ -1,0 +1,12 @@
+import { CharacterResponseModel } from '../models'
+
+export interface FilterCharactersByStatus {
+  perform(params: FilterCharactersByStatus.Params): Promise<FilterCharactersByStatus.Result>
+}
+
+export namespace FilterCharactersByStatus {
+  export type Params = {
+    status: string
+  }
+  export type Result = CharacterResponseModel
+}
