+++ rick-and-morty-app/src/domain/usecases/index.ts
@@ -1,5 +1,5 @@
 export * from './get-characters'
 export * from './get-character-by-id'
-export * from './filter-character-by-name'
-export * from './filter-character-by-species'
-export * from './filter-character-by-status'
+export * from './filter-characters-by-name'
+export * from './filter-characters-by-species'
+export * from './filter-characters-by-status'
